Fix issue with SCP 300 level causing errors
