+++ classes/class-ledyer-confirmation.php
@@ -79,6 +79,11 @@ class Confirmation {
 		}
 
 		$ledyer_order = ledyer()->api->get_order( $payment_id );
+		// If we could not get the Ledyer order, set it to an empty array and continue with the confirmation.
+		if ( is_wp_error( $ledyer_order ) ) {
+			$ledyer_order = array();
+		}
+
 		$order->set_transaction_id( $payment_id );
 
 		$ledyer_update_order_reference = ledyer()->api->update_order_reference( $payment_id, array( 'reference' => $order->get_order_number() ) );
@@ -89,7 +94,7 @@ class Confirmation {
 		}
 
 		$ledyer_note           = '';
-		$ledyer_payment_method = $ledyer_order['paymentMethod'];
+		$ledyer_payment_method = $ledyer_order['paymentMethod'] ?? array( 'provider' => '', 'type' => '' );
 		$ledyer_payment_status = ledyer()->api->get_payment_status( $payment_id );
 		if ( ! is_wp_error( $ledyer_payment_status ) ) {
 			$ledyer_payment_method = $ledyer_payment_status['paymentMethod'];
@@ -121,23 +126,7 @@ class Confirmation {
 
 		$order->set_payment_method_title( sprintf( '%s (Ledyer)', $method_title ) );
 
-		$note = sprintf(
-			// translators: 1: Payment ID, 2: Note from Ledyer.
-			__(
-				'New payment created in Ledyer with Payment ID %1$s. %2$s',
-				'ledyer-checkout-for-woocommerce'
-			),
-			$payment_id,
-			$ledyer_note
-		);
-		if ( 'advanceInvoice' === $ledyer_payment_type ) {
-			$order->update_status( 'on-hold', $note );
-		} elseif ( $order->needs_processing() ) {
-			$order->add_order_note( $note );
-			$order->payment_complete( $payment_id );
-		} else {
-			$order->update_status( 'on-hold', $note );
-		}
+		self::process_order_status( $ledyer_payment_status, $order, $payment_id );
 
 		$order->update_meta_data( '_ledyer_date_paid', gmdate( 'Y-m-d H:i:s' ) );
 		$order->save();
@@ -148,5 +137,118 @@ class Confirmation {
 		if ( is_wp_error( $response ) ) {
 			\Ledyer\Logger::log( "Couldn't acknowledge order {$payment_id}|{$order->get_id()}" );
 		}
+	}
+
+	/**
+	 * Process the order status based on the Ledyer payment status.
+	 *
+	 * @param array $ledyer_payment_status The Ledyer payment status response.
+	 * @param \WC_Order $order The WooCommerce order object.
+	 * @param string $ledyer_order_id The Ledyer order ID.
+	 *
+	 * @return void
+	 */
+	public static function process_order_status( $ledyer_payment_status, $order, $ledyer_order_id ) {
+		$ack_order = false;
+	    switch ( $ledyer_payment_status['status'] ) {
+	    	case \LedyerPaymentStatus::ORDER_PENDING:
+	    		if ( ! $order->has_status( array( 'on-hold', 'processing', 'completed' ) ) ) {
+	    			$note = sprintf(
+	    				__(
+	    					'New session created in Ledyer with Payment ID %1$s. %2$s',
+	    					'ledyer-checkout-for-woocommerce'
+	    				),
+	    				$ledyer_order_id,
+	    				$ledyer_payment_status['note']
+	    			);
+	    			$order->update_status( 'on-hold', $note );
+	    		}
+	    		break;
+	    	case \LedyerPaymentStatus::PAYMENT_PENDING:
+	    		if ( ! $order->has_status( array( 'on-hold', 'processing', 'completed' ) ) ) {
+	    			$note = sprintf(
+	    				__(
+	    					'New payment pending payment created in Ledyer with Payment ID %1$s. %2$s',
+	    					'ledyer-checkout-for-woocommerce'
+	    				),
+	    				$ledyer_order_id,
+	    				$ledyer_payment_status['note']
+	    			);
+	    			$order->update_status( 'on-hold', $note );
+	    			$ack_order = true;
+	    		}
+	    		break;
+	    	case \LedyerPaymentStatus::PAYMENT_CONFIRMED:
+	    		if ( ! $order->has_status( array( 'processing', 'completed' ) ) ) {
+	    			$note = sprintf(
+	    				__(
+	    					'Payment successfully confirmed in Ledyer with Payment ID %1$s. %2$s',
+	    					'ledyer-checkout-for-woocommerce'
+	    				),
+	    				$ledyer_order_id,
+	    				$ledyer_payment_status['note']
+	    			);
+
+
+	    			$order->add_order_note( $note );
+					$ack_order = true;
+
+					// If the order does not need payment, we need to ensure the "ready-for-capture" event has been triggered.
+	    			$ready_for_capture = $order->get_meta( '_ledyer_ready_for_capture', true );
+	    			if ( ! $order->needs_payment() && empty( $ready_for_capture ) ) {
+	    				// Set the metadata to indicate the order is still waiting for the ready for capture event.
+	    				$order->update_meta_data( '_ledyer_waiting_on_ready_for_capture', true );
+	    				$order->save();
+	    				Logger::log( "[SCHEDULER]: Order {$order->get_order_number()} is paid but waiting for ready_for_capture event." );
+	    				break;
+	    			}
+	    			$order->payment_complete( $ledyer_order_id );
+	    		}
+	    		break;
+	    	case \LedyerPaymentStatus::ORDER_CAPTURED:
+	    		$new_status = 'completed';
+
+	    		$settings = get_option( 'woocommerce_lco_settings' );
+
+	    		// Check if we should keep card payments in processing status.
+	    		if (
+	    			isset( $settings['keep_cards_processing'] )
+	    			&& 'yes' === $settings['keep_cards_processing']
+	    			&& isset( $ledyer_payment_status['paymentMethod'] )
+	    			&& isset( $ledyer_payment_status['paymentMethod']['type'] )
+	    			&& 'card' === $ledyer_payment_status['paymentMethod']['type']
+	    		) {
+	    			$new_status = 'processing';
+	    		}
+
+	    		$new_status = apply_filters( 'lco_captured_update_status', $new_status, $ledyer_payment_status );
+	    		$order->update_status( $new_status );
+	    		break;
+	    	case \LedyerPaymentStatus::ORDER_REFUNDED:
+	    		$order->update_status( 'refunded' );
+	    		break;
+	    	case \LedyerPaymentStatus::ORDER_CANCELLED:
+	    		$order->update_status( 'cancelled' );
+	    		break;
+	    }
+
+		// If we need to acknowledge the order, do it now.
+	    if ( $ack_order ) {
+	    	$response = ledyer()->api->acknowledge_order( $ledyer_order_id );
+	    	if ( is_wp_error( $response ) ) {
+	    		Logger::log( "[SCHEDULER]: Couldn't acknowledge order $ledyer_order_id" );
+	    		return;
+	    	}
+
+			// If the merchant reference was not already set, set it now.
+			$merchant_reference = $order->get_meta( '_ledyer_merchant_reference', true );
+	    	if ( empty( $merchant_reference ) ) {
+				$ledyer_update_order = ledyer()->api->update_order_reference( $ledyer_order_id, array( 'reference' => $order->get_order_number() ) );
+				if ( is_wp_error( $ledyer_update_order ) ) {
+					Logger::log( "[SCHEDULER]: Couldn't set merchant reference {$order->get_order_number()}" );
+					return;
+				}
+			}
+	    }
 	}
 }
